youtube-shorts-generator/requirements.txt
# Core (Phase 1)
PyYAML>=6.0
python-dotenv>=1.0.0

# External APIs
openai>=1.0.0
chromadb>=0.4.0
elevenlabs>=0.2.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Utilities
pydantic>=2.0.0
structlog>=23.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
